Deleting PostProcessingActivity redundation
