+++ app/src/main/AndroidManifest.xml
@@ -54,7 +54,6 @@
             android:name="com.yalantis.ucrop.UCropActivity"
             android:screenOrientation="portrait"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar" />
-        <activity android:name=".PostProcessingActivity"></activity>
     </application>
 
 </manifest>
