--- app/src/main/res/layout/activity_post_processing.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".PostProcessingActivity">
-
-    <com.yalantis.ucrop.view.UCropView
-        android:id="@+id/cropView"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_above="@id/layoutModifiers"/>
-
-    <LinearLayout
-        android:id="@+id/layoutModifiers"
-        android:orientation="vertical"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true">
-
-        <Button
-            android:id="@+id/btnConfirmCrop"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:text="Crop image"/>
-
-    </LinearLayout>
-
-</RelativeLayout>
