--- app/src/main/res/menu/crop_menu.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <item
-        android:id="@+id/btnResetChanges"
-        android:title="Reset"
-        android:icon="@drawable/ic_reset_changes"
-        app:showAsAction="always" />
-
-</menu>
